db/requirements.txt
python-dotenv
sqlparse
faker
mysql-connector-python
langchain
langchain-google-genai
langchain-openai
ollama
langchain-ollama
sqlglot
